Add a compression option
In order to target older operating systems, we need to be able to
set the compression to something that they'll suport, such as `xz`.
For example, these days, `zstd` is the default, but that won't work
on older systems.

This option allows you to specify the compression on the command line.
If you do not use the new setting, then the default value from `dpkg-deb`
will be used, just the same as before.